requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.3

# Config & Validation
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1

# Auth
PyJWT==2.8.0

# HTTP Client
httpx==0.27.2

# AI
langchain==0.3.1
langchain-groq==0.2.1
langchain-google-genai==2.0.7
langgraph==0.2.28
langchain-core==0.3.15
google-generativeai==0.8.3

# OpenTelemetry
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
